backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.15.0
sqlalchemy>=1.4.0,<2.0.0
alembic>=1.7.0,<2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
redis>=4.0.0
python-dotenv>=0.19.0
PyJWT>=2.4.0
requests>=2.28.0
cryptography>=3.4.8